Did some prep work to begin working on item functionality.
Decided that having resources for items was going to be too tedious and a pain. Could be wrong that maybe one resource could act for all items but I got the sense that I would need a resource for each and every item and that would just be dumb. Decided to go with a dictionary for all items and just call them with an ID num.

Also merged the chest dictionary and the newly created item dictionary into the global script because what else is the global script for if not to hold stuff like this. If they get too big or cumbersome I can always move them away. Try to keep tabs on how many places they get called from to make that change easier if it is done.